devices-emulator/pyproject.toml
[tool.poetry]
name = "devices-emulator"
version = "0.1.0"
description = ""
authors = ["fourvrd <kacperpiwowarski@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
